Add "shards edit" command
